kyc-python-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
easyocr==1.7.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-bidi==0.6.6
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
shapely==2.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
